cleanup some windows stuff
Window positions will change as layout gets settled on. Trying to
replicate what was originally there, I think the status and inventory
windows would overlap depending on commands ('b' backpack displays
inventory, otherwise the status window shows). Not the best design but
trying to just stick to what it was originally for now.